requirements/base.txt
# Core Django and REST
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-filter==23.3
django-rest-framework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.7

# Environment and utilities
python-dotenv==1.0.0
Pillow==10.0.1

# API Documentation
drf-yasg==1.21.7

# Date and time
python-dateutil==2.8.2

# Phone number validation
phonenumbers==8.13.20

# QR Code generation
qrcode==7.4.2

# PDF generation (optional for invoices)
reportlab==4.0.4

# Network Management
paramiko==3.4.0           # SSH for OLT/Huawei
librouteros==3.3.0        # Mikrotik API
netaddr==1.2.0            # IP address manipulation
requests==2.31.0          # HTTP requests for TR-069
